requirements.txt
anomalib==2.2.0
langchain==1.2.12
langchain-core==1.2.19
langchain-ollama==1.0.1
langchain-openai==1.1.11
lightning==2.6.0
matplotlib==3.10.8
numpy==1.26.4
ollama==0.6.1
openai==2.28.0
opencv-python==4.12.0.88
pandas==2.3.3
pillow==12.1.0
pydantic==2.12.5
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
